clippy: drop redundant usize cast in stellar tx.rs test
CI's `make check` step was failing at `lint` (Makefile:31) on
`clippy::unnecessary_cast` in patala-stellar/src/tx.rs:905 —
`tx.operations.len() as usize` casts a `usize` to `usize`, which a
newer clippy (CI's stable was 1.97.0 at the time of the failing run;
this repo's local toolchain is 1.94.0, which does not flag this
pattern) correctly rejects as dead code. Applied clippy's own suggested
fix verbatim: dropped the cast. `.len()` on stellar_xdr 22.2.0's `VecM`
does return `usize`, so this compiles and type-checks identically;
patala-stellar's test suite shows the same 84 passed / 0 failed /
3 ignored (unit) and 5 passed / 0 failed (doctests) before and after.

Agent-generated, machine-verified, NOT human-reviewed.

+++ patala-stellar/src/tx.rs
@@ -902,7 +902,7 @@ mod tests {
     fn the_builder_accepts_exactly_the_protocol_limit_and_refuses_one_more() {
         let ok: Vec<PaymentLeg> = (0..MAX_OPERATIONS).map(|i| leg(i as u8, 1)).collect();
         let tx = build_payment_transaction([1u8; 32], &ok, 1, 100, [0u8; 32]).unwrap();
-        assert_eq!(tx.operations.len() as usize, MAX_OPERATIONS);
+        assert_eq!(tx.operations.len(), MAX_OPERATIONS);
 
         let too_many: Vec<PaymentLeg> = (0..=MAX_OPERATIONS).map(|i| leg(i as u8, 1)).collect();
         let err = build_payment_transaction([1u8; 32], &too_many, 1, 100, [0u8; 32])
